+++ R/zzz.R
@@ -1,5 +1,9 @@
 .login_cache <- new.env(parent = emptyenv())
 
+.onLoad <- function(libname, pkgname) {
+  S7::methods_register()
+}
+
 .onAttach <- function(libname, pkgname) {
   packageStartupMessage(cli::format_inline(
     "{.pkg azr} {utils::packageVersion('azr')} | Azure OAuth 2.0 credential chain"